exploration: put containers in component's dumb index export
