+++ src/components/content/index.js
@@ -1,1 +1,14 @@
-export { default } from './container'
+import { connect } from 'react-redux'
+import { createStructuredSelector } from 'reselect'
+
+import { getDarkMode } from '../../store/ui'
+import { isLoggedIn } from '../../store/user'
+
+import Content from './content'
+
+const mapStateToProps = createStructuredSelector({
+  isLoggedIn,
+  darkMode: getDarkMode,
+})
+
+export default connect(mapStateToProps)(Content)
