--- src/components/layout/grid/container.js
@@ -1,11 +1,0 @@
-import { connect } from 'react-redux'
-
-import { getDarkMode } from '../../../store/ui'
-
-import Grid from './grid'
-
-const mapStateToProps = state => ({
-  darkMode: getDarkMode(state),
-})
-
-export default connect(mapStateToProps)(Grid)
