+++ src/components/sidebar/index.js
@@ -1,1 +1,12 @@
-export { default } from './container'
+import { connect } from 'react-redux'
+import { createStructuredSelector } from 'reselect'
+
+import { getDarkMode } from '../../store/ui'
+
+import Sidebar from './sidebar'
+
+const mapStateToProps = createStructuredSelector({
+  darkMode: getDarkMode,
+})
+
+export default connect(mapStateToProps)(Sidebar)
